Update context page values for all CTA events, reorganized enum values and added documentation
