+++ app/src/main/java/com/kickstarter/libs/AnalyticEvents.kt
@@ -15,12 +15,13 @@ import com.kickstarter.libs.utils.EventContextValues.CtaContextName.WATCH_PROJEC
 import com.kickstarter.libs.utils.EventContextValues.ContextTypeName.WATCH
 import com.kickstarter.libs.utils.EventContextValues.ContextTypeName.UNWATCH
 import com.kickstarter.libs.utils.EventContextValues.ContextTypeName.CREDIT_CARD
-import com.kickstarter.libs.utils.EventContextValues.PageViewedContextName.ADD_ONS
-import com.kickstarter.libs.utils.EventContextValues.PageViewedContextName.CHECKOUT
-import com.kickstarter.libs.utils.EventContextValues.PageViewedContextName.UPDATE_PLEDGE
-import com.kickstarter.libs.utils.EventContextValues.PageViewedContextName.REWARDS
-import com.kickstarter.libs.utils.EventContextValues.PageViewedContextName.PROJECT
-import com.kickstarter.libs.utils.EventContextValues.PageViewedContextName.THANKS
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.UPDATE_PLEDGE
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.ACTIVITY_FEED
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.ADD_ONS
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.CHECKOUT
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.REWARDS
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.PROJECT
+import com.kickstarter.libs.utils.EventContextValues.ContextPageName.THANKS
 import com.kickstarter.libs.utils.EventContextValues.CtaContextName.SEARCH
 import com.kickstarter.libs.utils.EventContextValues.DiscoveryContextType.PWL
 import com.kickstarter.libs.utils.EventContextValues.DiscoveryContextType.SUBCATEGORY_NAME
@@ -36,6 +37,7 @@ import com.kickstarter.libs.utils.ContextPropertyKeyName.CONTEXT_TYPE
 import com.kickstarter.libs.utils.ContextPropertyKeyName.CONTEXT_PAGE
 import com.kickstarter.libs.utils.ContextPropertyKeyName.CONTEXT_SECTION
 import com.kickstarter.libs.utils.ContextPropertyKeyName.CONTEXT_LOCATION
+import com.kickstarter.libs.utils.EventContextValues
 import com.kickstarter.libs.utils.EventContextValues.CtaContextName.DISCOVER
 import com.kickstarter.libs.utils.EventContextValues.LocationContextName.DISCOVER_ADVANCED
 import com.kickstarter.libs.utils.EventContextValues.LocationContextName.DISCOVER_OVERLAY
@@ -221,6 +223,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
     fun trackDiscoverProjectCTAClicked(){
         val props = hashMapOf<String, Any>()
         props[CONTEXT_CTA.contextName] = DISCOVER.contextName
+        props[CONTEXT_PAGE.contextName] = ACTIVITY_FEED.contextName
         props[CONTEXT_LOCATION.contextName]  = GLOBAL_NAV.contextName
         client.track(CTA_CLICKED.eventName, props)
     }
@@ -624,6 +627,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
     fun trackDiscoverSortCTA(discoveryParams: DiscoveryParams) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to DISCOVER_SORT.contextName)
         props[CONTEXT_LOCATION.contextName]  = DISCOVER_ADVANCED.contextName
+        props[CONTEXT_PAGE.contextName] = EventContextValues.ContextPageName.DISCOVER.contextName
         props.putAll(AnalyticEventsUtils.discoveryParamsProperties(discoveryParams))
         client.track(CTA_CLICKED.eventName, props)
     }
@@ -642,7 +646,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
     fun trackDiscoveryPageViewed(discoveryParams: DiscoveryParams) {
         val props = AnalyticEventsUtils.discoveryParamsProperties(discoveryParams).toMutableMap()
         props[DISCOVER_SORT.contextName] = discoveryParams.sort()?.name?.toLowerCase(Locale.ROOT) ?: ""
-        props[CONTEXT_PAGE.contextName] = DISCOVER.contextName
+        props[CONTEXT_PAGE.contextName] = EventContextValues.ContextPageName.DISCOVER.contextName
         client.track(PAGE_VIEWED.eventName, props)
     }
 
@@ -654,6 +658,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
     fun trackDiscoverFilterCTA(discoveryParams: DiscoveryParams) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to DISCOVER_FILTER.contextName)
         props[CONTEXT_LOCATION.contextName] = DISCOVER_OVERLAY.contextName
+        props[CONTEXT_PAGE.contextName] = EventContextValues.ContextPageName.DISCOVER.contextName
         props[CONTEXT_TYPE.contextName] = when {
             BooleanUtils.isTrue(discoveryParams.staffPicks()) -> PWL.contextName
             BooleanUtils.isTrue(discoveryParams.recommended()) -> RECOMMENDED.contextName
@@ -720,6 +725,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
      */
     fun trackSearchCTAButtonClicked(discoveryParams: DiscoveryParams) {
         val props = AnalyticEventsUtils.discoveryParamsProperties(discoveryParams).toMutableMap()
+        props[CONTEXT_PAGE.contextName] = EventContextValues.ContextPageName.DISCOVER.contextName
         props[CONTEXT_CTA.contextName] = SEARCH.contextName
         props[CONTEXT_LOCATION.contextName]  = GLOBAL_NAV.contextName
         client.track(CTA_CLICKED.eventName, props)
@@ -780,6 +786,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
     fun trackPledgeSubmitCTA(checkoutData: CheckoutData, pledgeData: PledgeData) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to PLEDGE_SUBMIT.contextName)
         props[CONTEXT_TYPE.contextName] = CREDIT_CARD.contextName
+        props[CONTEXT_PAGE.contextName] = CHECKOUT.contextName
         props.putAll(AnalyticEventsUtils.checkoutDataProperties(checkoutData, pledgeData, client.loggedInUser()))
         client.track(CTA_CLICKED.eventName, props)
     }
@@ -811,6 +818,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
      */
     fun trackPledgeInitiateCTA(projectData: ProjectData) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to PLEDGE_INITIATE.contextName)
+        props[CONTEXT_PAGE.contextName] = PROJECT.contextName
         props.putAll(AnalyticEventsUtils.projectProperties(projectData.project(), client.loggedInUser()))
         client.track(CTA_CLICKED.eventName, props)
     }
@@ -827,6 +835,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
      */
     fun trackSelectRewardCTA(pledgeData: PledgeData) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to REWARD_CONTINUE.contextName)
+        props[CONTEXT_PAGE.contextName] = REWARDS.contextName
         props.putAll(AnalyticEventsUtils.pledgeDataProperties(pledgeData, client.loggedInUser()))
         client.track(CTA_CLICKED.eventName, props)
     }
@@ -873,6 +882,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
     fun trackWatchProjectCTA(project: Project) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to WATCH_PROJECT.contextName)
         props[CONTEXT_TYPE.contextName] = if (project.isStarred) WATCH.contextName else UNWATCH.contextName
+        props[CONTEXT_PAGE.contextName] = PROJECT.contextName
         props.putAll(AnalyticEventsUtils.projectProperties(project, client.loggedInUser()))
         client.track(CTA_CLICKED.eventName, props)
     }
@@ -889,6 +899,7 @@ class AnalyticEvents(trackingClients: List<TrackingClientType?>) {
      */
     fun trackAddOnsContinueCTA(pledgeData: PledgeData) {
         val props: HashMap<String, Any> = hashMapOf(CONTEXT_CTA.contextName to ADD_ONS_CONTINUE.contextName)
+        props[CONTEXT_PAGE.contextName] = ADD_ONS.contextName
         props.putAll(AnalyticEventsUtils.pledgeDataProperties(pledgeData, client.loggedInUser()))
         client.track(CTA_CLICKED.eventName, props)
     }
