+++ app/src/main/java/com/kickstarter/viewmodels/ProfileViewModel.kt
@@ -168,7 +168,7 @@ interface ProfileViewModel {
 
             projectCardClicked
                     .compose(bindToLifecycle())
-                    .subscribe{ lake.trackProjectCardClicked(it, EventContextValues.PageViewedContextName.PROFILE.contextName) }
+                    .subscribe{ lake.trackProjectCardClicked(it, EventContextValues.ContextPageName.PROFILE.contextName) }
         }
 
         override fun emptyProfileViewHolderExploreProjectsClicked(viewHolder: EmptyProfileViewHolder) = this.exploreProjectsButtonClicked()
